third_party/WebKit/LayoutTests/fast/forms/month-multiple-fields/month-multiple-fields-ax-value-changed-notification-expected.txt

This test checks value changed accessibility notifications.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Focus Month=AXValueDescription: October
Focus Year=AXValueDescription: 2012
ValueChanged Year=AXValueDescription: 0005
ValueChanged Year=AXValueDescription: 0005
ValueChanged Year=AXValueDescription: 0005
ValueChanged Year=AXValueDescription: 0005

PASS successfullyParsed is true

TEST COMPLETE
